Really correct display of columns

+++ server/app/models.py
@@ -36,7 +36,7 @@ class Project(models.Model):
 class Column(models.Model):
     name = models.CharField(max_length=100, default="")
     wip = models.IntegerField()
-    location = models.IntegerField()
+    location = models.FloatField()
     parent_column = models.ForeignKey('self', null=True)
     board = models.ForeignKey(Board)
     is_border = models.BooleanField(default=False)
